+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog for Rapid Photo Downloader
 0.9.4 (2017-xx-xx)
 ------------------
 
+ - Add progress bars and badge counts when running under Ubuntu 17.10's Dash
+   to Dock extension.
+
  - Don't crash when locale is not correctly set.
 
 0.9.3 (2017-09-13)